changed variable kickingspeeds to floats

+++ src/utilities/Constants.cpp
@@ -264,24 +264,24 @@ std::map<int, bool> Constants::ROBOTS_WITH_KICKER() {
     return kickerRobots;
 }
 
-std::map<int, int> Constants::ROBOTS_MAXIMUM_KICK_TIME() {
-    static std::map<int, int> maximumKickTimes;
-    maximumKickTimes[0] = 25;
-    maximumKickTimes[1] = 31; // Tested: 6.1
-    maximumKickTimes[2] = 25; // Tested
-    maximumKickTimes[3] = 25;
-    maximumKickTimes[4] = 25;
-    maximumKickTimes[5] = 25; // Tested
-    maximumKickTimes[6] = 25;
-    maximumKickTimes[7] = 20; // Tested: 5.5
-    maximumKickTimes[8] = 40; // Tested: 5 is actualy 60
-    maximumKickTimes[9] = 25; // Tested: idk
-    maximumKickTimes[10] = 25;
-    maximumKickTimes[11] = 25;
-    maximumKickTimes[12] = 25;
-    maximumKickTimes[13] = 40; // Tested: 4.8 is actually 60
-    maximumKickTimes[14] = 25;
-    maximumKickTimes[15] = 25;
+std::map<int, float> Constants::ROBOTS_MAXIMUM_KICK_TIME() {
+    static std::map<int, float> maximumKickTimes;
+    maximumKickTimes[0] = 25.0;
+    maximumKickTimes[1] = 31.0; // Tested: 6.1
+    maximumKickTimes[2] = 25.0; // Tested
+    maximumKickTimes[3] = 25.0;
+    maximumKickTimes[4] = 25.0;
+    maximumKickTimes[5] = 25.0; // Tested
+    maximumKickTimes[6] = 25.0;
+    maximumKickTimes[7] = 20.0; // Tested: 5.5
+    maximumKickTimes[8] = 40.0; // Tested: 5 is actualy 60
+    maximumKickTimes[9] = 25.0; // Tested: idk
+    maximumKickTimes[10] = 25.0;
+    maximumKickTimes[11] = 25.0;
+    maximumKickTimes[12] = 25.0;
+    maximumKickTimes[13] = 40.0; // Tested: 4.8 is actually 60
+    maximumKickTimes[14] = 25.0;
+    maximumKickTimes[15] = 25.0;
 
     return maximumKickTimes;
 }
